+++ src/packages/Diagrams/src/Visuals/TableVisual/2Styles/Table.ts
