+++ src/packages/Diagrams/src/Visuals/TableVisual/2Types/tableVisualOptions.ts
