+++ src/packages/Diagrams/src/Visuals/TableVisual/2Utils/generateDefaultColumns.tsx
