+++ src/packages/Diagrams/src/Visuals/TableVisual/index.ts
@@ -1,3 +1,3 @@
-export * from './Hooks/useColumns';
+export * from './2Hooks/useColumns';
+export * from './2Types/tableVisualOptions';
 export * from './TableVisual';
-export * from './Types/tableVisualOptions';
