+++ src/packages/Diagrams/src/components/AnalyticsView.tsx
@@ -3,7 +3,6 @@ import { AnalyticsOptions, getSections } from '../types/analyticsOptions';
 import { Page, Wrapper } from './analyticsViewStyles';
 import { getChart } from './GetCharts';
 
-
 export interface AnalyticsViewProps<T> {
     data: T[];
     isLoading?: boolean;
